Backend/requirements.txt
# Backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0

# Database dependencies
sqlalchemy>=2.0.0
alembic>=1.12.0

# Authentication and security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Core ML dependencies
scikit-learn>=1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# HTTP requests
requests>=2.31.0
aiofiles>=23.2.1

# AI/ML for enhanced features
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
